backend/requirements.txt
# FastAPI & Server
fastapi==0.115.5
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Audio Processing
faster-whisper==1.0.3
pyannote.audio==4.0.2
torch>=2.0.0
torchaudio>=2.0.0
pydub==0.25.1
soundfile==0.12.1
librosa==0.10.2

# LLM Integration
openai==1.54.4
google-generativeai==0.8.3
anthropic==0.39.0
ollama==0.3.3

# Utilities
python-dotenv==1.0.1
numpy>=1.24.0
scipy>=1.11.0
pyyaml>=6.0.1

